requirements.txt
# Core
jinja2>=3.1.0
python-whois>=0.8.0

# PDF generation (optional but recommended)
weasyprint>=60.0

# Optional extras
requests>=2.31.0
